A lot of compiler L4 warnings fixed (some warnings disabled)
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@1982 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/parser/AviSplitter/AviReportWnd.cpp
@@ -262,7 +262,7 @@ bool CAviPlotterWnd::Create(CAviFile* pAF, CRect r, CWnd* pParentWnd)
 		DWORD* curchunks = DNew DWORD[pAF->m_avih.dwStreams];
 		memset(curchunks, 0, sizeof(DWORD)*pAF->m_avih.dwStreams);
 
-		CAviFile::strm_t::chunk2 cs2last = {-1, 0};
+		CAviFile::strm_t::chunk2 cs2last = {(DWORD)-1, 0};
 
 		while(1)
 		{
